Allow CSS custom properties (without checking)
This change is a temporary workaround that adds an interim/initial/partial level
of support for handling CSS custom properties and the `var(...)` function. The
intent is that we’ll add full support later, and then at that time back this
change out — or else at least just use it as a starting point for full support.

This change simply allows declarations to contain custom property names (that
is, identifiers starting with two dashes) and to contain the `var(...)` function
(with a custom property name as the first argument). Beyond that, it doesn’t add
any checking to report an error if a `var(...)` function contains a custom
property name that the stylesheet hasn’t actually defined, and no checking to
report an error if a particular property has an expression with a `var(...)`
function which resolves to a value that’s not allowed for that property.